# Stop precomputing chunk sizes