The non-concretely typed vector with chunks prevents constant propagation of the length of a static array. Alternative to https://github.com/JuliaDiff/ForwardDiff.jl/pull/707

This code was used to workaround https://github.com/JuliaLang/julia/issues/29887 but with modern Julia it just inhibits optimizations.