@KristofferC 2024-08-15
Or at least backport it