@devmotion 2024-08-15
I ran the benchmark in https://github.com/JuliaDiff/ForwardDiff.jl/pull/373, on the master branch and with this PR:

## master

```julia
julia> @btime ForwardDiff.gradient(rosenbrock, x);
  328.805 ns (3 allocations: 1.92 KiB)

julia> @btime ForwardDiff.gradient(rosenbrock, x);
  331.407 ns (3 allocations: 1.92 KiB)

julia> @btime ForwardDiff.gradient(rosenbrock, x);
  328.077 ns (3 allocations: 1.92 KiB)
```

## PR #708

```julia
julia> @btime ForwardDiff.gradient(rosenbrock, x);
  417.503 ns (3 allocations: 1.92 KiB)

julia> @btime ForwardDiff.gradient(rosenbrock, x);
  406.040 ns (3 allocations: 1.92 KiB)

julia> @btime ForwardDiff.gradient(rosenbrock, x);
  407.500 ns (3 allocations: 1.92 KiB)
```

## Setup

```julia
julia> versioninfo()
Julia Version 1.10.4
Commit 48d4fd48430 (2024-06-04 10:41 UTC)
Build Info:
  Official https://julialang.org/ release
Platform Info:
  OS: macOS (arm64-apple-darwin22.4.0)
  CPU: 10 × Apple M2 Pro
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-15.0.7 (ORCJIT, apple-m1)
Threads: 1 default, 0 interactive, 1 GC (on 6 virtual cores)
Environment:
  JULIA_PKG_USE_CLI_GIT = true
```