@KristofferC 2024-08-15
I'm not getting so obvious differences:

```
(@v1.10) pkg> st ForwardDiff
Status `~/.julia/environments/v1.10/Project.toml`
  [f6369f11] ForwardDiff v0.11.0-DEV `https://github.com/JuliaDiff/ForwardDiff.jl.git#master`

julia> @btime ForwardDiff.gradient(rosenbrock, x)
  401.741 ns (3 allocations: 1.92 KiB)

---------

shell> git status
On branch kc/no_prec_chunk
Your branch is up to date with 'origin/kc/no_prec_chunk'.

nothing to commit, working tree clean

julia> @btime ForwardDiff.gradient(rosenbrock, x)
  405.835 ns (3 allocations: 1.92 KiB)
```