@KristofferC 2024-08-15
This was an easier way too see the regression:

```
julia> const x = rand(10)

# master
julia> @btime ForwardDiff.Chunk($x)
  3.792 ns (0 allocations: 0 bytes)
ForwardDiff.Chunk{10}()

# PR
julia> @btime ForwardDiff.Chunk($x)
  55.894 ns (0 allocations: 0 bytes)
ForwardDiff.Chunk{10}()
```

I made a new commit to try work around that.
```