@devmotion 2024-08-15
I reran the benchmark #373 with the latest commit:

```julia
julia> @btime ForwardDiff.gradient(rosenbrock, x);
  358.333 ns (3 allocations: 1.92 KiB)

julia> @btime ForwardDiff.gradient(rosenbrock, x);
  346.847 ns (3 allocations: 1.92 KiB)

julia> @btime ForwardDiff.gradient(rosenbrock, x);
  347.417 ns (3 allocations: 1.92 KiB)
```

And that's what I get when I benchmark `Chunk`:

```julia
# master
julia> @btime ForwardDiff.Chunk($x)
  3.458 ns (0 allocations: 0 bytes)

# PR
julia> @btime ForwardDiff.Chunk($x)
  3.625 ns (0 allocations: 0 bytes)
```