@KristofferC 2024-08-19
So would you say this is ok as it is now?