@devmotion 2024-08-19
Personally, I think the benefits (for type inference and static arrays) are worth the ~5% regression in these benchmarks, even more so since these benchmarks only focus on the construction of `Chunk` and `GradientConfig`.